Vaihingen dataset - Get a main label for patch
This program is now able to read the satellite data patches, get the
pixels values, create pixels labels and get a main label for each patch.
The patches' main labels will only be useful for a regular CNN made with
Keras.
The final goal is to create a conv-deconv neural network. We will need
the pixel level labels in this case.